Move to AssertingWeight
