+++ lucene/core/src/test/org/apache/lucene/search/TestSortOptimization.java
@@ -16,13 +16,6 @@
  */
 package org.apache.lucene.search;
 
-import static org.apache.lucene.search.SortField.FIELD_DOC;
-import static org.apache.lucene.search.SortField.FIELD_SCORE;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.FloatDocValuesField;
@@ -37,12 +30,18 @@ import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
-import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.RandomIndexWriter;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.store.Directory;
-import org.apache.lucene.util.Bits;
 import org.apache.lucene.util.LuceneTestCase;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import static org.apache.lucene.search.SortField.FIELD_DOC;
+import static org.apache.lucene.search.SortField.FIELD_SCORE;
 
 public class TestSortOptimization extends LuceneTestCase {
 
@@ -60,7 +59,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     final IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     final SortField sortField = new SortField("my_field", SortField.Type.LONG);
     final Sort sort = new Sort(sortField);
     final int numHits = 3;
@@ -145,7 +144,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     final IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     final SortField sortField = new SortField("my_field", SortField.Type.LONG);
     final Sort sort = new Sort(sortField);
     final int numHits = 3;
@@ -184,7 +183,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     final IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     final int numHits = 3;
     final int totalHitsThreshold = 3;
 
@@ -236,7 +235,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     final IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     final int numHits = 3;
     final int totalHitsThreshold = 3;
 
@@ -315,7 +314,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     final IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     final SortField sortField = new SortField("my_field", SortField.Type.FLOAT);
     final Sort sort = new Sort(sortField);
     final int numHits = 3;
@@ -663,7 +662,7 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     SortField sortField = new SortField("my_field", SortField.Type.LONG);
     TopFieldDocs topDocs =
         searcher.search(new MatchAllDocsQuery(), 1 + random().nextInt(100), new Sort(sortField));
@@ -708,7 +707,7 @@ public class TestSortOptimization extends LuceneTestCase {
     seqNos.sort(Long::compare);
     IndexReader reader = DirectoryReader.open(writer);
     writer.close();
-    IndexSearcher searcher = newIndexSearcher(reader);
+    IndexSearcher searcher = newSearcher(reader);
     SortField sortField = new SortField("seq_no", SortField.Type.LONG);
     int visitedHits = 0;
     ScoreDoc after = null;
@@ -731,53 +730,5 @@ public class TestSortOptimization extends LuceneTestCase {
     }
     reader.close();
     dir.close();
-  }
-
-  private static class ChunkedBulkScorer extends BulkScorer {
-    private final BulkScorer in;
-
-    ChunkedBulkScorer(BulkScorer in) {
-      this.in = in;
-    }
-
-    @Override
-    public int score(LeafCollector collector, Bits acceptDocs, int min, int max)
-        throws IOException {
-      while (min < max) {
-        final int interval;
-        if (random().nextInt(100) <= 5) {
-          interval = 1 + random().nextInt(10);
-        } else {
-          interval = 1 + random().nextInt(random().nextBoolean() ? 100 : 5000);
-        }
-        final int newMax = Math.toIntExact(Math.min(min + interval, max));
-        min = in.score(collector, acceptDocs, min, newMax);
-      }
-      return min;
-    }
-
-    @Override
-    public long cost() {
-      return in.cost();
-    }
-  }
-
-  private static IndexSearcher newIndexSearcher(IndexReader reader) {
-    return new IndexSearcher(reader) {
-      @Override
-      protected void search(List<LeafReaderContext> leaves, Weight weight, Collector collector)
-          throws IOException {
-        if (random().nextBoolean()) {
-          weight =
-              new FilterWeight(weight) {
-                @Override
-                public BulkScorer bulkScorer(LeafReaderContext context) throws IOException {
-                  return new ChunkedBulkScorer(super.bulkScorer(context));
-                }
-              };
-        }
-        super.search(leaves, weight, collector);
-      }
-    };
   }
 }
